LayoutTests/imported/w3c/web-platform-tests/css/css-break/page-break-legacy-shorthands-expected.txt


FAIL Legacy values of the shorthands work as expected assert_equals: expected "left" but got "always"
FAIL New values work on the new longhands, but serialize to the empty string in the legacy shorthands assert_equals: expected "" but got "auto"
PASS New values of the break longhands don't work on legacy shorthands
FAIL Legacy shorthands really never appear on cssText, even when there are variable references assert_equals: expected -1 but got 0
